requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
fastapi==0.115.11
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
mysql-connector-python==9.2.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.3.5
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0